added GameContents post component
game contents post consists of the headline text (e.g. kim is winning!), details of the sport, and availability of the match (public/private), match date, which teams are playing and the score, and footer buttons Like Comment Share, and footer buttons Join and Add. The GameContents component is currently not added into newsfeed page yet, but was in for testing/styling purposes.